requirements.txt
langchain-ollama==0.2.0
python-dotenv==1.0.1
langchain==0.3.3
langchain-community==0.3.2
langgraph==0.2.37
